Huge commit: -update copyrights to 2013 -rebranding to "The Land of Alembrum" -start working on soundFX replacements (20% done) -minor bugs fixed (soundFX happening twice in dialog phase)

+++ zildo/src/zildo/monde/sprites/persos/ia/BasicMover.java
@@ -10,8 +10,8 @@ public class BasicMover extends Mover {
 	
 	@Override
 	protected void move() {
-		int px = (int) Math.signum( ((int) target.x - mobile.x));
-		int py = (int) Math.signum( ((int) target.y - mobile.y));
+		int px = (int) Math.signum( ( target.x - mobile.x));
+		int py = (int) Math.signum( ( target.y - mobile.y));
 		
 		mobile.x+=px;
 		mobile.y+=py;
